# Skip with statements automatically where needed