While the `with` context managers are very useful, it is annoying to have to remember the `if xyz.opened` and analogous, and also the double indentation can get in the way. The solution outlined towards the end of #255 would solve this but it is not that ideal since it still requires the user to remember to raise the exception, and also it is worse in terms of readability.

I discovered a [project](https://github.com/justengel/conditional_context) that allowed skipping `with` blocks using dark magic (I'm not 100% sure on what exactly it does, but my best understanding is that it modifies the system trace to make python think the exception was raised just inside the `with` block, and then discards the exception in `__exit__`) and tried to adapt it to wrap an existing context manager.

This PR implements `imgui._skipwith.SkipWith(context, should_skip)`, an internal class used to wrap an existing imgui context manager (first argument) and optionally skip the `with` block body based on a condition (second argument). When applied to the relevant context managers this is the result:

Continues to work:
```python
imgui.begin("Popup Example")
if imgui.button("Open Modal Popup"):
    imgui.open_popup("Modal Popup")
if imgui.begin_popup_modal("Modal Popup")[0]:
    imgui.text("Select an option:")
    imgui.separator()
    imgui.selectable("Red Pill")
    imgui.selectable("Blue Pill")
    imgui.end_popup()
imgui.end()
```
Also continues to work (behavior of #255):
```python
with imgui.begin("Popup Example"):
    if imgui.button("Open Modal Popup"):
        imgui.open_popup("Modal Popup")
    with imgui.begin_popup_modal("Modal Popup") as popup:
        if popup.opened:
            # Double indentation, additional if check
            imgui.text("Select an option:")
            imgui.separator()
            imgui.selectable("Red Pill")
            imgui.selectable("Blue Pill")
```
Can be rewritten as:
```python
with imgui.begin("Popup Example"):
    if imgui.button("Open Modal Popup"):
        imgui.open_popup("Modal Popup")
    with imgui.begin_popup_modal("Modal Popup"):
        # No need to check popup.opened
        # This 'with' block gets skipped if opened is False
        imgui.text("Select an option:")
        imgui.separator()
        imgui.selectable("Red Pill")
        imgui.selectable("Blue Pill")
```