@Willy-JL 2022-03-28
Nevermind, I hadn't run all the tests...
Also in hindsight it's a bad idea to block all other user code in a with block on the library's side...