@KinoxKlark 2022-03-29
I think that this is an interesting idea and we shouldn't be so prompt in rejecting it. Don't worry about failing tests, 2.0 is still in dev, and breaking stuff is expected going to this new major release. Moreover, as far as I can see, tests that broke are those introduced in the `with` functionality which may need to be adapted with your PR.

I am a little bit more concerned with performances, it seems to me that this adds a lot of work on top of simple function calls. However, this is already the case for the `with` functionality and I am far from having a deep understanding of those things in python, so it may not be a problem. Maybe we should test those things.

> Also in hindsight it's a bad idea to block all other user code in a with block on the library's side...

Can you develop? I am not sure to understand all implications of your proposal, I'll have to dig deeper into it to grasp a better understanding.

Anyway, thanks for the help :)