@Willy-JL 2022-03-29
I'm glad you also think this is an interesting idea. My main concern, that I only realized after submitting the PR, is that this simply cuts off all the content of the with block, meaning that if the user has some custom logic inside it, it will be completely ignored. Perhaps we could make this feature togglable? Some food for thought...

Also about the tests, what put me off is that after realizing that above, then the tests that failed where mainly about catching exceptions inside the modified with statements, and since I already understand little about how the method to skip the with blocks works, I was intimidated to look into the issue and gave up a bit too quickly.

However I'm willing to look more into since you also think this might be a good idea, so let me know about the first concern with skipping custom user logic.
(Also since I deleted the branch on my fork I might have to make a new PR when it comes to implementing this, if we get there. We can use this issue to discuss in the meantime)