@Willy-JL 2022-03-29
I was thinking with my OOP brain and not with a logical train of thought... a wrapper class is completely overkill for this purpose, simple functions do the job just fine.

This way the logic for the `with` statements remains largely untouched, when using the 'old' style with `if` checks there are no issues with the class type, and now the `.opened` `AttributeErrors` are fixed. Down from 14 failed tests to 3, and those 3 all involve raising exceptions inside the `with` statement, will have to look deeper into it for those.