@Willy-JL 2022-03-29
Performance wise I get fps between high 550s and low 560s fps with if-style code, while in the mid 550s using the with-style code. Looks to me like this `with skip` functionality does impact performance but in a veeery negligible manner.

Also about the 3 failing tests it's obvious: these tests check if an exception can be thrown from inside the `with` statement, but the exception will never be thrown in this case because the `with` block is skipped altogether (when `.opened` is `False`, and this is the case with the tests because there is no user to interact with the widgets). So these 3 tests will need to be reconsidered if we decide to go forward with this feature.

So the last concern that remains in my opinion is the same I had originally:
 - should it be always enabled?
 - should it be a module level toggle (`imgui.skip_with_statements(True/False)`)?
 - should it be a begin level toggle (`imgui.begin("Example", skip_with=True/False)`)?
 - if togglable, should it be enabled or disabled by default?
 - should it even be part of the library?

I'm open to any kind of suggestion!