@KinoxKlark 2022-03-30
Thanks for the clarification and proposal!

> My main concern, that I only realized after submitting the PR, is that this simply cuts off all the content of the with block, meaning that if the user has some custom logic inside it, it will be completely ignored.

Ok, I understand it now.

> [...] simple functions do the job just fine

Indeed, this looks cleaner and seems to fix the previously raised issue. However I have some doubts about the `global _orig_trace`, it may only be a lack of understanding on my part, but wouldn't this break for nested skipped-with ? I may have to look deeper into it to be sure. Anyway, it starts to look like something which is great!

---

Concerning perfs, I am not sure that looking at FPS count is a great way to measure it. This is influenced by too many other things and I think that perf problems will show only for big projects which extensively use the functionality, i.e., with many calls by frame.

In my opinion, a better measurement technic is to isolate as much as possible some examples of use in small functions (similarly to test cases) and use `timeit` to compute the average computation time on 1000000 calls. I made a quick and dirty setup to do it and I measure taking your popup example:

```
Branch: dev/version-2.0
old style 'begin/end': 0.7116959999999999 µs
old style 'with': 0.9190690000000004 µs
```

```
Branch: skip-with-statements
old style 'begin/end': 0.7256831999999998 µs
old style 'with': 0.9228938000000007 µs
new 'withskip': 1.5783046999999994 µs
```

(I also quickly tested other types of begin/end)
(I didn't check if the output is correct in all cases)

This isn't by any means an extensive measurement and may need more rigor, but we can already note two things:
* Your change doesn't seems to impact the already existing setup which is good
* New `withskip` looks 1.5x slower than `with` and 2x slower than `begin/end` (note that `with` is also ~1.5x slower than `begin/end`)

This may be accelerated using cython instead of python, I'll need to dig deeper into that. For now, I am not sure if this behavior should be the default one. It is clearly more programmer-friendly but the penalty feels a bit harsh for such a used feature of the lib. Note that this also raises concerns with regard to the original `with` proposal. I will have to run some experiments and think about all of that.

Yes, it could be a togglable feature but wouldn't it also cost a set of conditional around each use of the feature? And wouldn't it be a bit confusing at use? I still think the whole thing is a good idea and I would prefer to have it as default or not have it at all, but we need to figure out if we can make it as efficient as possible. This pattern is used everywhere in imgui and if we don't take special care with it it will clumps up and bite us in the back.

*Note: We not only need to consider the raw timing of function calls but also how many times we expect those calls to occur in a "normal" application in order to estimate the total penalty.*

*Note: I would be curious to compare perf on an old branch without the `with` functionality. Comparisons with `begin/end` may be biased...*